+++ vm/src/nova_vm.rs
@@ -12,7 +12,7 @@ use move_deps::{
     },
     move_bytecode_utils::Modules,
     move_core_types::{
-        account_address::AccountAddress, effects::ChangeSet, language_storage::TypeTag,
+        account_address::AccountAddress, effects::ChangeSet,
         vm_status::StatusCode,
     },
     move_vm_runtime::{
@@ -23,12 +23,11 @@ use move_deps::{
     move_vm_types::gas::UnmeteredGasMeter,
 };
 use crate::{
-    natives::table::{NativeTableContext, TableChangeSet, TableHandle, TableResolver}, 
+    natives::table::{NativeTableContext, TableChangeSet, TableResolver}, 
     args_validator::check_args_address, 
-    size_change_set::{SizeDelta, SizeChangeSet, AccountSizeChangeSet}
+    size_change_set::AccountSizeChangeSet
 };
 use std::{
-    borrow::Borrow,
     collections::{BTreeMap, BTreeSet},
     sync::Arc,
 };
@@ -47,8 +46,8 @@ use crate::{
         nova_natives,
     },
     session::{SessionExt, SessionOutput},
-    storage::data_view_resolver::{StoredSizeResolver, TableMetaResolver},
-    table_meta::{find_all_address_occur, TableMetaChangeSet, TableOwnerDataCache},
+    storage::data_view_resolver::{StoredSizeResolver},
+    table_meta::{resolve_table_ownership, resolve_table_size_change_by_account , TableMetaChangeSet, TableOwnerDataCache},
     NovaVMError,
 };
 
@@ -309,7 +308,7 @@ impl NovaVM {
         let temporary_session = self.create_session(remote_cache, session_id);
         let mut data_cache = TableOwnerDataCache::new(remote_cache);
 
-        self.resolve_table_ownership(
+        resolve_table_ownership(
                 temporary_session,
                 &session_output.0,
                 &session_output.2, 
@@ -320,7 +319,7 @@ impl NovaVM {
                 e.into_vm_status()
             })?;
 
-        let accounts_table_size_changes = self.resolve_table_size_change_by_account(
+        let accounts_table_size_changes = resolve_table_size_change_by_account(
             &session_output.2, 
             &session_output.4,
             &mut data_cache   
@@ -495,124 +494,6 @@ impl NovaVM {
         PartialVMError::new(StatusCode::CONSTRAINT_NOT_SATISFIED)
             .with_message(format!("metadata and code bundle mismatch: {}", msg))
             .finish(Location::Undefined)
-    }
-
-    fn resolve_table_ownership<
-        S: TableMetaResolver + MoveResolver + TableResolver + StoredSizeResolver,
-    >(
-        &self,
-        session: SessionExt<'_, '_, S>,
-        change_set: &ChangeSet,
-        table_change_set: &TableChangeSet,
-        table_owner_data_cache: &mut TableOwnerDataCache<'_, S>,
-    ) -> VMResult<()> {
-
-        let new_tables = &table_change_set.borrow().new_tables;
-
-        println!("new table {:?}", table_change_set.new_tables);
-        println!("remove table {:?}", table_change_set.removed_tables);
-
-        for i in table_change_set.removed_tables.iter() {
-            table_owner_data_cache.del_owner(i);
-            table_owner_data_cache.del_size(i);
-        }
-
-        for (addr, account_change_set) in change_set.borrow().accounts().iter() {
-            println!("checking changeset for {}", addr);
-            for (i, op) in account_change_set.resources().iter() {
-                let ty_tag = TypeTag::Struct(i.clone());
-                let ty_layout = session.get_type_layout(&ty_tag)?;
-                let res = find_all_address_occur(op, &ty_layout)?;
-
-                for address_found in res.into_iter() {
-                    let found_handle = TableHandle(address_found);
-                    // address is new table's handle or
-                    // already stored table's handle
-
-                    if new_tables.contains_key(&found_handle) || table_owner_data_cache.is_registerd_table(&found_handle)? {
-                        table_owner_data_cache.set_owner(&found_handle, *addr);
-                    }
-                }
-            }
-        }
-
-        for (outer_handle, change) in table_change_set.changes.iter() {
-            for (_key, op) in &change.entries {
-                let res = find_all_address_occur(op,  &change.value_layout)?;
-
-                for address_found in res.iter() {
-                    let found_handle = TableHandle(*address_found);
-                    if new_tables.contains_key(&found_handle) || table_owner_data_cache.is_registerd_table(&found_handle)? {
-                        table_owner_data_cache.set_owner(&found_handle, outer_handle.0);
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    fn resolve_table_size_change_by_account<
-        S: TableMetaResolver + MoveResolver + TableResolver + StoredSizeResolver,
-    >(
-        &self,
-        table_change_set: &TableChangeSet,
-        table_size_change: &SizeChangeSet<TableHandle>,
-        // table_meta_change_set: &TableMetaChangeSet,
-        table_owner_data_cache: &mut TableOwnerDataCache<'_, S>,
-    ) -> VMResult<AccountSizeChangeSet> {
-        let mut accounts_delta: AccountSizeChangeSet = AccountSizeChangeSet::default();
-
-
-        println!("new table {:?}", table_change_set.new_tables);
-        println!("remove table {:?}", table_change_set.removed_tables);
-
-        // handle previous size
-        for (handle, owner) in table_owner_data_cache.get_owner_changes().into_iter() {
-            let size = match table_owner_data_cache.get_size(handle)? {
-                Some(s) => s,
-                None => {continue;},    // skip new table
-            };
-            
-            match owner {
-                Some(new_owner) => {
-                        let old_owner = table_owner_data_cache.get_root_owner(handle)?.unwrap();
-                        accounts_delta.move_size(old_owner, new_owner.clone(), size);
-                    },
-                None => {
-                    let acc = table_owner_data_cache.get_root_owner(handle)?.unwrap();
-                    let delta = SizeDelta::decreasing(size);
-                    accounts_delta.insert_size(acc, delta);
-                },
-            }
-        }
-        println!("table size owning???? {:?}", accounts_delta);
-
-        
-        // handle size changes
-        for (handle, size_delta) in table_size_change.changes() {
-            // todo: should check deleted table?? yes
-            if table_change_set.removed_tables.contains(handle) {
-                continue;
-            }
-            let owner = table_owner_data_cache.get_root_owner(handle)?.unwrap();
-            
-            accounts_delta.insert_size(owner, size_delta.clone());
-
-            let current_size = table_owner_data_cache.get_size(&handle)?.unwrap_or(0);
-
-            let mut delta = SizeDelta::increasing( current_size);
-            delta.merge(size_delta.clone());
-            if delta.is_decrease {
-                panic!("dsdads");
-            }
-            let new_size = delta.amount;
-        
-            table_owner_data_cache.set_size(&handle, new_size);
-        }
-
-        println!("{:?}", accounts_delta);
-
-        Ok(accounts_delta)
     }
 }
