+++ vm/src/storage/data_view_resolver.rs
@@ -11,7 +11,7 @@ use log::error;
 use move_deps::move_binary_format::errors::{Location, PartialVMError, VMError, VMResult};
 use move_deps::move_core_types::{
     account_address::AccountAddress,
-    language_storage::{ModuleId, StructTag, TypeTag},
+    language_storage::{ModuleId, StructTag},
     resolver::{ModuleResolver, ResourceResolver},
     vm_status::StatusCode,
 };
